- backend: autocomplete: use RecTryLock to find out if we can obtain shutdown_mutex instead of exception throw

+++ backend/wbpublic/sqlide/autocomplete_object_name_cache.cpp
@@ -257,28 +257,27 @@ std::vector<std::string> AutoCompleteCache::get_matching_objects(const std::stri
 {
   if (!_shutdown)
   {
-    try
+    // Ensures shutdown is not done while processing
+    base::RecMutexTryLock sd_lock(_shutdown_mutex);
+    if (!sd_lock.locked())
     {
-      // Ensures shutdown is not done while processing
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_sqconn_mutex, true);
-      sqlite::query q(*_sqconn, "SELECT name FROM " + cache + " WHERE name LIKE ? ESCAPE '\\'");
-      q.bind(1, base::escape_sql_string(prefix, true) + "%");
-      if (q.emit())
-      {
-        std::vector<std::string> items;
-        boost::shared_ptr<sqlite::result> matches(q.get_result());
-        do
-        {
-          items.push_back(matches->get_string(0));
-        } while (matches->next_row());
-
-        return items;
-      }
+      log_debug3("Can't obtain lock on shutdown_mutex for get_matching_objects call.\n");
+      return std::vector<std::string>();
     }
-    catch (std::exception &exc)
+
+    base::RecMutexLock lock(_sqconn_mutex);
+    sqlite::query q(*_sqconn, "SELECT name FROM " + cache + " WHERE name LIKE ? ESCAPE '\\'");
+    q.bind(1, base::escape_sql_string(prefix, true) + "%");
+    if (q.emit())
     {
-      log_error("Exception caught while retrieving objects for caches %s with prefix %s :%s", cache.c_str(), prefix.c_str(), exc.what());
+      std::vector<std::string> items;
+      boost::shared_ptr<sqlite::result> matches(q.get_result());
+      do
+      {
+        items.push_back(matches->get_string(0));
+      } while (matches->next_row());
+
+      return items;
     }
   }
 
@@ -295,30 +294,29 @@ std::vector<std::string> AutoCompleteCache::get_matching_objects(const std::stri
 {
   if (!_shutdown)
   {
-    try
+    // Ensures shutdown is not done while processing
+    base::RecMutexTryLock sd_lock(_shutdown_mutex);
+    if (!sd_lock.locked())
     {
-      // Ensures shutdown is not done while processing
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_sqconn_mutex, true);
-      sqlite::query q(*_sqconn, "SELECT name FROM " + cache + " WHERE schema_id LIKE ? ESCAPE '\\' "
-        "AND name LIKE ? ESCAPE '\\'");
-      q.bind(1, schema.empty() ? "%" : base::escape_sql_string(schema, true));
-      q.bind(2, base::escape_sql_string(prefix, true) + "%");
-      if (q.emit())
-      {
-        std::vector<std::string> items;
-        boost::shared_ptr<sqlite::result> matches(q.get_result());
-        do
-        {
-          items.push_back(matches->get_string(0));
-        } while (matches->next_row());
-
-        return items;
-      }
+      log_debug3("Can't obtain lock on shutdown_mutex for get_matching_objects call.\n");
+      return std::vector<std::string>();
     }
-    catch (std::exception &exc)
+
+    base::RecMutexLock lock(_sqconn_mutex);
+    sqlite::query q(*_sqconn, "SELECT name FROM " + cache + " WHERE schema_id LIKE ? ESCAPE '\\' "
+      "AND name LIKE ? ESCAPE '\\'");
+    q.bind(1, schema.empty() ? "%" : base::escape_sql_string(schema, true));
+    q.bind(2, base::escape_sql_string(prefix, true) + "%");
+    if (q.emit())
     {
-      log_error("Exception caught while retrieving objects for caches %s with prefix %s for schema %s :%s", cache.c_str(), prefix.c_str(), schema.c_str(), exc.what());
+      std::vector<std::string> items;
+      boost::shared_ptr<sqlite::result> matches(q.get_result());
+      do
+      {
+        items.push_back(matches->get_string(0));
+      } while (matches->next_row());
+
+      return items;
     }
   }
 
@@ -335,31 +333,30 @@ std::vector<std::string> AutoCompleteCache::get_matching_objects(const std::stri
 {
   if (!_shutdown)
   {
-    try
+    // Ensures shutdown is not done while processing
+    base::RecMutexTryLock sd_lock(_shutdown_mutex);
+    if (!sd_lock.locked())
     {
-      // Ensures shutdown is not done while processing
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_sqconn_mutex, true);
-      sqlite::query q(*_sqconn, "SELECT name FROM " + cache + " WHERE schema_id LIKE ? ESCAPE '\\' "
-        "AND table_id LIKE ? ESCAPE '\\' AND name LIKE ? ESCAPE '\\'");
-      q.bind(1, schema.empty() ? "%" : base::escape_sql_string(schema, true));
-      q.bind(2, table.empty() ? "%" : base::escape_sql_string(table, true));
-      q.bind(3, base::escape_sql_string(prefix, true) + "%");
-      if (q.emit())
-      {
-        std::vector<std::string> items;
-        boost::shared_ptr<sqlite::result> matches(q.get_result());
-        do
-        {
-          items.push_back(matches->get_string(0));
-        } while (matches->next_row());
-
-        return items;
-      }
+      log_debug3("Can't obtain lock on shutdown_mutex for get_matching_objects call.\n");
+      return std::vector<std::string>();
     }
-    catch (std::exception &exc)
+
+    base::RecMutexLock lock(_sqconn_mutex);
+    sqlite::query q(*_sqconn, "SELECT name FROM " + cache + " WHERE schema_id LIKE ? ESCAPE '\\' "
+      "AND table_id LIKE ? ESCAPE '\\' AND name LIKE ? ESCAPE '\\'");
+    q.bind(1, schema.empty() ? "%" : base::escape_sql_string(schema, true));
+    q.bind(2, table.empty() ? "%" : base::escape_sql_string(table, true));
+    q.bind(3, base::escape_sql_string(prefix, true) + "%");
+    if (q.emit())
     {
-      log_error("Exception caught while retrieving objects for caches %s with prefix %s for schema %s.%s :%s", cache.c_str(), prefix.c_str(), schema.c_str(), table.c_str(), exc.what());
+      std::vector<std::string> items;
+      boost::shared_ptr<sqlite::result> matches(q.get_result());
+      do
+      {
+        items.push_back(matches->get_string(0));
+      } while (matches->next_row());
+
+      return items;
     }
   }
 
@@ -389,28 +386,28 @@ bool AutoCompleteCache::refresh_schema_cache_if_needed(const std::string &schema
 
   if (!_shutdown)
   {
-    try
+    // Ensures shutdown is not done while processing
+    base::RecMutexTryLock sd_lock(_shutdown_mutex);
+    if (!sd_lock.locked())
     {
-      // Ensures shutdown is not done while processing
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_sqconn_mutex, true);
-      sqlite::query q(*_sqconn, "SELECT last_refresh FROM schemas WHERE name LIKE ? ESCAPE '\\' ");
-      q.bind(1, schema.empty() ? "%" : base::escape_sql_string(schema, true));
-      if (q.emit())
+      log_debug3("Can't obtain lock on shutdown_mutex for refresh_schema_cache_if_needed call.\n");
+      return false;
+    }
+
+    base::RecMutexLock lock(_sqconn_mutex);
+    sqlite::query q(*_sqconn, "SELECT last_refresh FROM schemas WHERE name LIKE ? ESCAPE '\\' ");
+    q.bind(1, schema.empty() ? "%" : base::escape_sql_string(schema, true));
+    if (q.emit())
+    {
+      boost::shared_ptr<sqlite::result> matches(q.get_result());
+      // If a value is set for last_refresh then schema info is already loaded in cache.
+      if (matches->get_int(0) != 0)
       {
-        boost::shared_ptr<sqlite::result> matches(q.get_result());
-        // If a value is set for last_refresh then schema info is already loaded in cache.
-        if (matches->get_int(0) != 0)
-        {
-          log_debug3("schema %s is already cached\n", schema.c_str());
-          return false;
-        }
+        log_debug3("schema %s is already cached\n", schema.c_str());
+        return false;
       }
     }
-    catch (std::exception &exc)
-    {
-      log_error("Exception caught while trying to refresh cache for schema: %s:%s", schema.c_str(), exc.what());
-    }
+
   }
 
   // Add tasks to load various schema objects. They will then update the last_refresh value.
@@ -529,8 +526,6 @@ void AutoCompleteCache::refresh_cache_thread()
   }
 
   // Signal the main thread that the worker thread is (about to be) gone.
-//  _refresh_thread = NULL;
-
   _cache_working.post();
 
   if (_feedback && !_shutdown)
@@ -1073,8 +1068,13 @@ void AutoCompleteCache::update_schemas(const std::vector<std::string> &schemas)
     if (!_shutdown)
     {
       // Ensures shutdown is not done while processing
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_sqconn_mutex, true);
+      base::RecMutexTryLock sd_lock(_shutdown_mutex);
+      if (!sd_lock.locked())
+      {
+        log_debug3("Can't obtain lock on shutdown_mutex for update_schemas call.\n");
+        return;
+      }
+      base::RecMutexLock lock(_sqconn_mutex);
  
       std::map<std::string, int> old_schema_update_times;
       {
@@ -1165,8 +1165,13 @@ void AutoCompleteCache::update_object_names(const std::string &cache, const std:
   {
     if (!_shutdown)
     {
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_sqconn_mutex, true);
+      base::RecMutexTryLock sd_lock(_shutdown_mutex);
+      if (!sd_lock.locked())
+      {
+        log_debug3("Can't obtain lock on shutdown_mutex for update_object_names call.\n");
+        return;
+      }
+      base::RecMutexLock lock(_sqconn_mutex);
       sqlide::Sqlite_transaction_guarder trans(_sqconn, false);
       {
         sqlite::execute del(*_sqconn, "delete from " + cache);
@@ -1203,8 +1208,13 @@ void AutoCompleteCache::update_object_names(const std::string &cache, const std:
     if (!_shutdown)
     {
       // Ensures shutdown is not done while processing
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_sqconn_mutex, true);
+      base::RecMutexTryLock sd_lock(_shutdown_mutex);
+      if (sd_lock.locked())
+      {
+        log_debug3("Can't obtain lock on shutdown_mutex for update_object_names call.\n");
+         return;
+      }
+      base::RecMutexLock lock(_sqconn_mutex);
 
       sqlide::Sqlite_transaction_guarder trans(_sqconn, false); // Will be committed when we go out of the scope.
 
@@ -1240,8 +1250,13 @@ void AutoCompleteCache::update_object_names(const std::string &cache, const std:
     try
     {
       // Ensures shutdown is not done while processing
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_sqconn_mutex, true);
+      base::RecMutexTryLock sd_lock(_shutdown_mutex);
+      if (!sd_lock.locked())
+      {
+        log_debug3("Can't obtain lock on shutdown_mutex for update_object_names call.\n");
+        return;
+      }
+      base::RecMutexLock lock(_sqconn_mutex);
 
       sqlide::Sqlite_transaction_guarder trans(_sqconn, false);
 
@@ -1279,51 +1294,50 @@ void AutoCompleteCache::add_pending_refresh(RefreshTask::RefreshType type, const
 {
   if (!_shutdown)
   {
-    try
+
+    base::RecMutexTryLock sd_lock(_shutdown_mutex);
+    if (!sd_lock.locked())
     {
-    base::RecMutexLock sd_lock(_shutdown_mutex, true);
-    base::RecMutexLock lock(_pending_mutex, true);
+      log_debug3("Can't obtain lock on shutdown_mutex for add_pending_refresh call.\n");
+      return;
+    }
+    base::RecMutexLock lock(_pending_mutex);
 
-      // Add the new task only if there isn't already one of the same type and for the same objects.
-      bool found = false;
-      for (std::list<RefreshTask>::const_iterator i = _pending_tasks.begin(); !found && i != _pending_tasks.end(); ++i)
-      {
-        if (i->type != type)
-          continue;
+    // Add the new task only if there isn't already one of the same type and for the same objects.
+    bool found = false;
+    for (std::list<RefreshTask>::const_iterator i = _pending_tasks.begin(); !found && i != _pending_tasks.end(); ++i)
+    {
+      if (i->type != type)
+        continue;
 
-        switch (type) {
-          case RefreshTask::RefreshSchemas:
-          case RefreshTask::RefreshVariables:
-          case RefreshTask::RefreshEngines:
-          case RefreshTask::RefreshUDFs:
-            found = true;
-            break;
+      switch (type) {
+        case RefreshTask::RefreshSchemas:
+        case RefreshTask::RefreshVariables:
+        case RefreshTask::RefreshEngines:
+        case RefreshTask::RefreshUDFs:
+          found = true;
+          break;
 
-          case RefreshTask::RefreshTables:
-          case RefreshTask::RefreshViews:
-          case RefreshTask::RefreshProcedures:
-          case RefreshTask::RefreshFunctions:
-            found = i->schema_name == schema;
-            break;
+        case RefreshTask::RefreshTables:
+        case RefreshTask::RefreshViews:
+        case RefreshTask::RefreshProcedures:
+        case RefreshTask::RefreshFunctions:
+          found = i->schema_name == schema;
+          break;
 
-          case RefreshTask::RefreshTriggers:
-          case RefreshTask::RefreshColumns:
-          case RefreshTask::RefreshLogfileGroups:
-          case RefreshTask::RefreshTableSpaces:
-            found = (i->schema_name == schema) && (i->table_name == table);
-            break;
-        }
-        if (found)
+        case RefreshTask::RefreshTriggers:
+        case RefreshTask::RefreshColumns:
+        case RefreshTask::RefreshLogfileGroups:
+        case RefreshTask::RefreshTableSpaces:
+          found = (i->schema_name == schema) && (i->table_name == table);
           break;
       }
+      if (found)
+        break;
+    }
 
-      if (!found)
-        _pending_tasks.push_back(RefreshTask(type, schema, table));
-    }
-    catch (std::exception &exc)
-    {
-      log_error("Exception caught while adding refresh call for %s.%s: %s\n", schema.c_str(), table.c_str(), exc.what());
-    }
+    if (!found)
+      _pending_tasks.push_back(RefreshTask(type, schema, table));
   }
 
   // Create the worker thread if there's work to do. Does nothing if there's already a thread.
@@ -1339,23 +1353,23 @@ bool AutoCompleteCache::get_pending_refresh(RefreshTask &task)
 
   if (!_shutdown)
   {
-    try
-    {
-      // Ensures shutdown is not done while processing
-      base::RecMutexLock sd_lock(_shutdown_mutex, true);
-      base::RecMutexLock lock(_pending_mutex, true);
 
-      if (!_pending_tasks.empty())
-      {
-        ret_val = true;
-        task = _pending_tasks.front();
-        _pending_tasks.pop_front();
-      }
-    }
-    catch (std::exception &exc)
+    // Ensures shutdown is not done while processing
+    base::RecMutexTryLock sd_lock(_shutdown_mutex);
+    if (!sd_lock.locked())
     {
-      log_error("Exception caught while getting pending refresh %s\n",  exc.what());
+      log_debug3("Can't obtain lock on shutdown_mutex for get_pending_refresh call.\n");
+      return false;
     }
+    base::RecMutexLock lock(_pending_mutex);
+
+    if (!_pending_tasks.empty())
+    {
+      ret_val = true;
+      task = _pending_tasks.front();
+      _pending_tasks.pop_front();
+    }
+
   }
 
   return ret_val;
